+++ src/views/system/dept/dialog.vue
@@ -1,7 +1,7 @@
 <template>
-	<div class="system-add-dept-container">
-		<el-dialog title="新增部门" v-model="state.isShowDialog" width="769px">
-			<el-form :model="state.ruleForm" size="default" label-width="90px">
+	<div class="system-dept-dialog-container">
+		<el-dialog :title="state.dialog.title" v-model="state.dialog.isShowDialog" width="769px">
+			<el-form ref="deptDialogFormRef" :model="state.ruleForm" size="default" label-width="90px">
 				<el-row :gutter="35">
 					<el-col :xs="24" :sm="24" :md="24" :lg="24" :xl="24" class="mb20">
 						<el-form-item label="上级部门">
@@ -60,21 +60,24 @@
 			<template #footer>
 				<span class="dialog-footer">
 					<el-button @click="onCancel" size="default">取 消</el-button>
-					<el-button type="primary" @click="onSubmit" size="default">新 增</el-button>
+					<el-button type="primary" @click="onSubmit" size="default">{{ state.dialog.submitTxt }}</el-button>
 				</span>
 			</template>
 		</el-dialog>
 	</div>
 </template>
 
-<script setup lang="ts" name="systemAddDept">
-import { reactive, onMounted } from 'vue';
+<script setup lang="ts" name="systemDeptDialog">
+import { reactive, ref } from 'vue';
+
+// 定义子组件向父组件传值/事件
+const emit = defineEmits(['refresh']);
 
 // 定义变量内容
+const deptDialogFormRef = ref();
 const state = reactive({
-	isShowDialog: false,
 	ruleForm: {
-		deptLevel: [], // 上级部门
+		deptLevel: [] as string[], // 上级部门
 		deptName: '', // 部门名称
 		person: '', // 负责人
 		phone: '', // 手机号
@@ -84,26 +87,51 @@ const state = reactive({
 		describe: '', // 部门描述
 	},
 	deptData: [] as DeptTreeType[], // 部门数据
+	dialog: {
+		isShowDialog: false,
+		type: '',
+		title: '',
+		submitTxt: '',
+	},
 });
 
 // 打开弹窗
-const openDialog = () => {
-	state.isShowDialog = true;
+const openDialog = (type: string, row: RowDeptType) => {
+	if (type === 'edit') {
+		row.deptLevel = ['vueNextAdmin'];
+		row.person = 'lyt';
+		row.phone = '12345678910';
+		row.email = 'vueNextAdmin@123.com';
+		state.ruleForm = row;
+		state.dialog.title = '修改部门';
+		state.dialog.submitTxt = '修 改';
+	} else {
+		state.dialog.title = '新增部门';
+		state.dialog.submitTxt = '新 增';
+		// 清空表单，此项需加表单验证才能使用
+		// nextTick(() => {
+		// 	deptDialogFormRef.value.resetFields();
+		// });
+	}
+	state.dialog.isShowDialog = true;
+	getMenuData();
 };
 // 关闭弹窗
 const closeDialog = () => {
-	state.isShowDialog = false;
+	state.dialog.isShowDialog = false;
 };
 // 取消
 const onCancel = () => {
 	closeDialog();
 };
-// 新增
+// 提交
 const onSubmit = () => {
 	closeDialog();
+	emit('refresh');
+	// if (state.dialog.type === 'add') { }
 };
 // 初始化部门数据
-const initTableData = () => {
+const getMenuData = () => {
 	state.deptData.push({
 		deptName: 'vueNextAdmin',
 		createTime: new Date().toLocaleString(),
@@ -131,10 +159,6 @@ const initTableData = () => {
 		],
 	});
 };
-// 页面加载时
-onMounted(() => {
-	initTableData();
-});
 
 // 暴露变量
 defineExpose({