+++ src/views/system/user/index.vue
@@ -9,14 +9,14 @@
 					</el-icon>
 					查询
 				</el-button>
-				<el-button size="default" type="success" class="ml10" @click="onOpenAddUser">
+				<el-button size="default" type="success" class="ml10" @click="onOpenAddUser('add')">
 					<el-icon>
 						<ele-FolderAdd />
 					</el-icon>
 					新增用户
 				</el-button>
 			</div>
-			<el-table :data="state.tableData.data" style="width: 100%">
+			<el-table :data="state.tableData.data" v-loading="state.tableData.loading" style="width: 100%">
 				<el-table-column type="index" label="序号" width="60" />
 				<el-table-column prop="userName" label="账户名称" show-overflow-tooltip></el-table-column>
 				<el-table-column prop="userNickname" label="用户昵称" show-overflow-tooltip></el-table-column>
@@ -34,7 +34,9 @@
 				<el-table-column prop="createTime" label="创建时间" show-overflow-tooltip></el-table-column>
 				<el-table-column label="操作" width="100">
 					<template #default="scope">
-						<el-button :disabled="scope.row.userName === 'admin'" size="small" text type="primary" @click="onOpenEditUser(scope.row)">修改</el-button>
+						<el-button :disabled="scope.row.userName === 'admin'" size="small" text type="primary" @click="onOpenEditUser('edit', scope.row)"
+							>修改</el-button
+						>
 						<el-button :disabled="scope.row.userName === 'admin'" size="small" text type="primary" @click="onRowDel(scope.row)">删除</el-button>
 					</template>
 				</el-table-column>
@@ -53,8 +55,7 @@
 			>
 			</el-pagination>
 		</el-card>
-		<AddUer ref="addUserRef" />
-		<EditUser ref="editUserRef" />
+		<UserDialog ref="userDialogRef" @refresh="getTableData()" />
 	</div>
 </template>
 
@@ -63,12 +64,10 @@ import { defineAsyncComponent, reactive, onMounted, ref } from 'vue';
 import { ElMessageBox, ElMessage } from 'element-plus';
 
 // 引入组件
-const AddUer = defineAsyncComponent(() => import('/@/views/system/user/component/addUser.vue'));
-const EditUser = defineAsyncComponent(() => import('/@/views/system/user/component/editUser.vue'));
+const UserDialog = defineAsyncComponent(() => import('/@/views/system/user/dialog.vue'));
 
 // 定义变量内容
-const addUserRef = ref();
-const editUserRef = ref();
+const userDialogRef = ref();
 const state = reactive<SysUserState>({
 	tableData: {
 		data: [],
@@ -82,7 +81,8 @@ const state = reactive<SysUserState>({
 });
 
 // 初始化表格数据
-const initTableData = () => {
+const getTableData = () => {
+	state.tableData.loading = true;
 	const data = [];
 	for (let i = 0; i < 2; i++) {
 		data.push({
@@ -102,14 +102,17 @@ const initTableData = () => {
 	}
 	state.tableData.data = data;
 	state.tableData.total = state.tableData.data.length;
+	setTimeout(() => {
+		state.tableData.loading = false;
+	}, 500);
 };
 // 打开新增用户弹窗
-const onOpenAddUser = () => {
-	addUserRef.value.openDialog();
+const onOpenAddUser = (type: string) => {
+	userDialogRef.value.openDialog(type);
 };
 // 打开修改用户弹窗
-const onOpenEditUser = (row: RowUserType) => {
-	editUserRef.value.openDialog(row);
+const onOpenEditUser = (type: string, row: RowUserType) => {
+	userDialogRef.value.openDialog(type, row);
 };
 // 删除用户
 const onRowDel = (row: RowUserType) => {
@@ -119,6 +122,7 @@ const onRowDel = (row: RowUserType) => {
 		type: 'warning',
 	})
 		.then(() => {
+			getTableData();
 			ElMessage.success('删除成功');
 		})
 		.catch(() => {});
@@ -126,14 +130,16 @@ const onRowDel = (row: RowUserType) => {
 // 分页改变
 const onHandleSizeChange = (val: number) => {
 	state.tableData.param.pageSize = val;
+	getTableData();
 };
 // 分页改变
 const onHandleCurrentChange = (val: number) => {
 	state.tableData.param.pageNum = val;
+	getTableData();
 };
 // 页面加载时
 onMounted(() => {
-	initTableData();
+	getTableData();
 });
 </script>